chore: Refactor move service to encapsulate getById* methods
- Add defaultInclude to move service

- Add _getById method to move service

- #getById and #getByIdWithEvents
  - Remove all params except id
  - Explicitly pass desired includes

- add setMoveWithEvents middleware
  - loads move using #getByIdWithEvents
  - inflates unpopulated resources / process timeline events

- moves shared code between move view and timeline to `app/move/controllers/view/view.locals.js`
