+++ app/move/controllers/timeline.test.js
@@ -1,6 +1,5 @@
 const proxyquire = require('proxyquire')
 
-const populateResources = sinon.stub()
 const presenters = {
   moveToMetaListComponent: sinon.stub().returns('moveToMetaListComponentArgs'),
   assessmentToTagList: sinon.stub().returns('assessmentToTagListArgs'),
@@ -8,15 +7,16 @@ const presenters = {
     .stub()
     .returns('eventsToTimelineComponentArgs'),
 }
-const moveService = {
-  getById: sinon.stub().resolves({ id: 'moveId' }),
-}
+
 const getTabsUrls = sinon.stub().returns('tab_urls')
 
+const getViewLocals = sinon.stub().returns({
+  locals: 'view.locals',
+})
+
 const timelineController = proxyquire('./timeline', {
-  '../../../common/lib/populate-resources': populateResources,
   '../../../common/presenters': presenters,
-  '../../../common/services/move': moveService,
+  './view/view.locals': getViewLocals,
   './view/view.tabs.urls': getTabsUrls,
 })
 
@@ -37,14 +37,10 @@ describe('Move controllers', function () {
       presenters.moveToMetaListComponent.resetHistory()
       presenters.assessmentToTagList.resetHistory()
       presenters.eventsToTimelineComponent.resetHistory()
-      moveService.getById.resetHistory()
-      moveService.getById.resolves(mockMove)
-      populateResources.resetHistory()
+      getViewLocals.resetHistory()
       getTabsUrls.resetHistory()
       req = {
-        params: {
-          id: 'moveId',
-        },
+        move: mockMove,
       }
     })
 
@@ -57,43 +53,13 @@ describe('Move controllers', function () {
         expect(getTabsUrls).to.be.calledOnceWithExactly(mockMove)
       })
 
-      it('should fetch timeline events for move', function () {
-        expect(moveService.getById).to.be.calledOnceWithExactly(mockMove.id, {
-          include: [
-            'profile',
-            'profile.person',
-            'from_location',
-            'to_location',
-            'timeline_events',
-            'timeline_events.eventable',
-          ],
-          populateResources: true,
-        })
-      })
-
-      it('should populate resources', function () {
-        expect(populateResources).to.be.calledOnceWithExactly(
-          mockMove.timeline_events
-        )
-      })
-
       it('should transform the data for presentation', function () {
-        expect(presenters.moveToMetaListComponent).to.be.calledOnceWithExactly(
-          mockMove
-        )
-        expect(presenters.assessmentToTagList).to.be.calledOnceWithExactly(
-          mockMove.profile.assessment_answers
-        )
-        expect(
-          presenters.eventsToTimelineComponent
-        ).to.be.calledOnceWithExactly(mockMove.timeline_events, mockMove)
+        expect(getViewLocals).to.be.calledOnceWithExactly(req)
       })
 
       it('should render the timeline', function () {
         expect(res.render).to.be.calledOnceWithExactly('move/views/timeline', {
-          move: mockMove,
-          moveSummary: 'moveToMetaListComponentArgs',
-          tagList: 'assessmentToTagListArgs',
+          locals: 'view.locals',
           timeline: 'eventsToTimelineComponentArgs',
           urls: {
             tabs: 'tab_urls',
